Replace macro pointer access in 'mp_module___main__'
Cannot use pointer access in initializer lists. Instead we delcare a global 'mp_obj_dict_t' type object that represents the active value (depending on what context you're in)

Untested - in a previous attempt 'const' was removed from the structure and the '.globals' member was changed directly. I hope that leaving 'const' in this declaration will not cause trouble beause the .globals member is not defined as constant and is a pointer so it should be evaluated when needed, thereby catching changes to 'mp_active_dict_main'