Update state access to go through the active context node
Creating a context_node type will allow the user to create a linked list of contexts that can be swapped in and out of the 'active' position. Adding more to the context is then as easy as modifying the structure in the type definition

A 'void mp_context_switch(mp_context_node_t* node)' function is added to handle switching.

In port main files it is used with the 'mp_context_head' pointer as the argument because the head is set to point at a statically allocated state that produces very similar behavior to the master branch of MicroPython