Set main task id manually
Normally the ID for a given context is set when the task is created - by using ```mp_task_register```. In the case of the main task (the REPL task) the context and state structure are statically allocated and are the first entry of the ```mp_context_head``` list head. Since they are statically allocated we need to manually set the ID when we find out what the taskID will be.

For new (dynamically allocated) contexts just use ```mp_task_register``` as shown in ```modmultipython.c```