Remove 'MP_CONTEXT_DEFAULT_ID'
Using a constant like that clashes with the concept of using a unique value of the running OS task to identify the MicroPython state. For example when switching back to ```mp_task``` the context switch function tries to find a context with an ID that matches the xTaskHandle of ```mp_task```. However for the main (statically allocated) task **if** we used an arbitrary number then this search would fail and the state would now be inconsistent.