Simplify MultiPython module interface
Generally speaking with MultiPython you can:
* start
* stop
* suspend
* resume
* get info

for processes that exist. Except for 'start' the operations default to applying to all contexts. If you specify a taskID, or list of taskIDs, then the operation is only applied to those contexts (if they exist). If you specify 'None' then the operation is only performed on the current task/context.

The start interface is not done. For now you pass in a string and optionally an integer. The string will represent the path to the file to execute unless the second argument is provided and equal to 0 in which case the string will be interpreted as valid MicroPython syntax