Pin main REPL task to a single core
This is a short-term solution to the REPL autocomplete breaking after some time. See, I noticed that occasionally autocomplete (tab complete) would break after doing a few 'import' statements. After a little rubber-ducking I thought that since the REPL task had no affinity for a particular core it might not always be accessing the proper mirror variable for the 'mp_module___main__' (because access is determined by which core the calling code is running on)

So, pinning the REPL to one task ensures that it is always accessing the same mirror of 'mp_module___main__' from which the autocomplete is determined.

It is important to note that this problem *should* only affect modules/functions that need to access these mirror variables:
* mp_obj_list_t mp_active_sys_path_objs[MICROPY_NUM_CORES]
* mp_obj_list_t mp_active_sys_argv_objs[MICROPY_NUM_CORES]
* mp_obj_dict_t mp_active_loaded_modules_dicts[MICROPY_NUM_CORES]
* mp_obj_dict_t mp_active_dict_mains[MICROPY_NUM_CORES]

Which affects primarily: modsys (for path and argv) and autocomplete.

It *should* be OK to create new processes that do not have core affinity because all access to state variables is conditionally controlled by the core that called the access.