Add first shot at 'addressable' module
This module is ultimately the heart and soul of Mach1. It gives you fully dynamic manipulation of LED outputs via micropython.

Controllers represent the physical outputs for LED data - currently there are two, one for the built-in status led and one for the Addressable LED port

Each controller contains a linked list of fixtures, all of which must operate on the same protocol (e.g. apa102 vs ws2812). Fixtures can be dynamically added, removed, or swapped in the order.

Output is scheduled by the controller's timer, which can also by dynamically changed (period, start/stop, etc). The timer is an esp timer which allows good resolution w/ minimal overhead. (and it is independent for each output)

Fixtures have a 'set' (name subject to change) function that allows setting a contiguous run of LEDs colors beginning at a particular index.

This is lightly tested on the status led, but I still need to make sure it works on the main output (ALED)