Debugged ALED output, rough draft
Previously ALED output wasn't working. Now it is.

Added zeroing of the memory when a chain is recomputed, as well as application of the protocol's masking rules (this allows fixtures to retain their position, b/c otherwise it sends all zeros which is not recognized by APA102 as a valid LED.)

Added offset of ```fixture.set()``` data by the specified index number, so that LEDs go in the right place. Also added a check to make sure that LED data does not write over into adjacent fixtures.

Next steps:
* major cleanup of module structure. e.g. why do you start a controller through the addressable module directly (addressable.start(controller_id)) when you could use a controller method? (e.g. controller_obj.start()).
* add [a] way(s) to get a fixture from a controller, b/c right now creating additional fixtures in multipython processes just appends them to the chain - not what we wanted! (I mean, we want that too, but mostly just to get existing fixtures)
* add information getters to all relevant things
* add output timer control
* add additional output control (controller.start / controller.stop / controller.delete [used to free up memory])
* I'm sure there is more... its a work in progress