Add (primitive) support for external SPI RAM on Wrover modules
Using the proper sdkconfig file (sdkconfig.spiram_malloc_caps) allows us to put all the micropython heaps in the SPI RAM, leaving internal RAM for the important stuff like DMA buffers, bluetooth, etc...

One big learning experience - for the longest time I was stumped as to why the ```heap_caps_malloc(..., MALLOC_CAP_SPIRAM)``` calls were failing when trying to start a new multipython process. Well - I had left the 'main.c' heap allocation method untouched and it called for taking the largest possible block of memory - anywhere. So of course the fix was as simple as giving the main task some sane amount of memory. (Hadcoded)

Future work includes allowing the user to specify (request) the memory size for a new process, and ultimately to be able to on-the-fly increase memory usage for tasks.

If you have trouble with SPI RAM messages on a WROOM just make sure your SDKCONFIG file is set properly in the makefile.