Begin support for OTA updates
So far things are OK-ish.

We've successfully connected to reverb.echoictech.com to access /ota/ota.txt and /ota/application.bin

When trying to write that file into a partition the ESP32 aborts because it cannot find the ota data. I suspect (from initial googling) that this is because MicroPython, not using the typical ESP-IDF build system, does not create a bootloader that respects the partition settings from the sdkconfig file.

I need to consider at least these two options moving forward:
1. Get the current setup to respect the IDF's bootloader creation process
1. Create a version of Micro/MultiPython that works more seamlessly with the IDF (like the Loboris repo...). At the same time I would try to keep the structure of MicroPython identical so that it is easier to merge updates from them in the future... (visual image of creating a tight rope between the Twin Towers, where one tower is MicroPython and the other is the ESP-IDF) (because I finally think I have enough understanding of creating build processes!) (And it would be really nice to get to use the ESP-IDF menuconfig - while still getting improvements from MicroPython...)