BT SPP control of uPy REPL
We're now hooked in so that the SPP can send and receive commands/output from the uPy REPL. It's not working flawlessly yet (bytes seem to be dropped when sending data back to the phone/computer) but it's a start. Also might be worth trying to use BLE instead of (or in addition to?) BT Classic because iOS does not support SPP for standard developers.