Change 'main.py' to use scripts stored in flash memory
In this example the 'main.py' script starts up files called 'scr_0_brakelight.py' and 'scr_1_signals.py'

Considering that brake lights and turn signals are important functions we wouldn't want to allow a failure of the SD card (connection, or not present) to prevent them from operating. So these files were copied from the SD card to the flash memory area. This area is much less likely to fail than the SD card, at least while the microcontroller is operating normally.